DB connection bug fix
mongo-atlas needs an IP address whitelist for an app to connect to the database
since we don't have a guaranteed range of IP addresses that the heroku app will
run on, we can't make a reliable whitelist for it.

The best solution is to use a heroku addon and I chose mlab so that this problem
will be handled by heroku.